cmds update

+++ minishell/parser/ft_atoi.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: asabbar <asabbar@student.42.fr>            +#+  +:+       +#+        */
+/*   By: zait-sli <zait-sli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/25 14:46:11 by asabbar           #+#    #+#             */
-/*   Updated: 2022/06/02 13:57:50 by asabbar          ###   ########.fr       */
+/*   Updated: 2022/06/09 02:18:19 by zait-sli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parser.h"
 
-long long	ft_atoi(const char *str)
+long double	ft_atoi(const char *str)
 {
 	size_t	i;
 	size_t	sign;
-	size_t	result;
+	long double	result;
 
 	i = 0;
 	sign = 1;
